python_suites/infotainment_python_suite/requirements.txt
python-can>=4.3.1
cantools>=39.4.4
pytest>=7.4.0
pytest-timeout>=2.2.0
